requirements.txt
# 基础依赖
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
requests==2.32.5
urllib3==2.6.2

# 数据库
psycopg2-binary==2.9.11
# 如果你的 Importer 或 FastAPI 使用了 ORM，请务必添加下行
# SQLAlchemy==2.0.25 

# 配置与环境 (推荐)
python-dotenv==1.0.1